update to 10.4.6 - 1 January 2022 snapshot (LTS)

+++ src/timers.c
@@ -1,5 +1,5 @@
 /*
- * FreeRTOS Kernel V10.4.4
+ * FreeRTOS Kernel V10.4.6
  * Copyright (C) 2021 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
  *
  * SPDX-License-Identifier: MIT
@@ -57,8 +57,8 @@
 #if ( configUSE_TIMERS == 1 )
 
 /* Misc definitions. */
-    #define tmrNO_DELAY                      ( ( TickType_t ) 0U )
-    #define tmrMAX_TIME_BEFORE_OVERFLOW      ( ( TickType_t ) -1 )
+    #define tmrNO_DELAY                    ( ( TickType_t ) 0U )
+    #define tmrMAX_TIME_BEFORE_OVERFLOW    ( ( TickType_t ) -1 )
 
 /* The name assigned to the timer service task.  This can be overridden by
  * defining trmTIMER_SERVICE_TASK_NAME in FreeRTOSConfig.h. */
@@ -359,11 +359,11 @@
         configASSERT( ( xTimerPeriodInTicks > 0 ) );
 
         /* Ensure the infrastructure used by the timer service task has been
-            * created/initialised. */
+         * created/initialised. */
         prvCheckForValidListAndQueue();
 
         /* Initialise the timer structure members using the function
-            * parameters. */
+         * parameters. */
         pxNewTimer->pcTimerName = pcTimerName;
         pxNewTimer->xTimerPeriodInTicks = xTimerPeriodInTicks;
         pxNewTimer->pvTimerID = pvTimerID;
@@ -458,7 +458,7 @@
             }
             else
             {
-                pxTimer->ucStatus &= ~tmrSTATUS_IS_AUTORELOAD;
+                pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_AUTORELOAD );
             }
         }
         taskEXIT_CRITICAL();
@@ -517,7 +517,7 @@
         /* Insert the timer into the appropriate list for the next expiry time.
          * If the next expiry time has already passed, advance the expiry time,
          * call the callback function, and try again. */
-        while ( prvInsertTimerInActiveList( pxTimer, ( xExpiredTime + pxTimer->xTimerPeriodInTicks ), xTimeNow, xExpiredTime ) != pdFALSE )
+        while( prvInsertTimerInActiveList( pxTimer, ( xExpiredTime + pxTimer->xTimerPeriodInTicks ), xTimeNow, xExpiredTime ) != pdFALSE )
         {
             /* Advance the expiry time. */
             xExpiredTime += pxTimer->xTimerPeriodInTicks;
@@ -536,6 +536,7 @@
 
         /* Remove the timer from the list of active timers.  A check has already
          * been performed to ensure the list is not empty. */
+
         ( void ) uxListRemove( &( pxTimer->xTimerListItem ) );
 
         /* If the timer is an auto-reload timer then calculate the next
@@ -546,7 +547,7 @@
         }
         else
         {
-            pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+            pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
         }
 
         /* Call the timer callback. */
@@ -825,7 +826,7 @@
                             }
                             else
                             {
-                                pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                             }
 
                             /* Call the timer callback. */
@@ -842,7 +843,7 @@
                     case tmrCOMMAND_STOP:
                     case tmrCOMMAND_STOP_FROM_ISR:
                         /* The timer has already been removed from the active list. */
-                        pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                        pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                         break;
 
                     case tmrCOMMAND_CHANGE_PERIOD:
@@ -872,7 +873,7 @@
                                 }
                                 else
                                 {
-                                    pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                    pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                                 }
                             }
                         #else /* if ( configSUPPORT_DYNAMIC_ALLOCATION == 1 ) */
@@ -881,7 +882,7 @@
                                  * could not have been dynamically allocated. So there is
                                  * no need to free the memory - just mark the timer as
                                  * "not active". */
-                                pxTimer->ucStatus &= ~tmrSTATUS_IS_ACTIVE;
+                                pxTimer->ucStatus &= ( ( uint8_t ) ~tmrSTATUS_IS_ACTIVE );
                             }
                         #endif /* configSUPPORT_DYNAMIC_ALLOCATION */
                         break;
